finagle css properties into one place
